refactor: rename edit folder to fields folder
